[ThroughputScorer] Simplify last-selected bonus
Change the way that a last-selected network bonus is applied.
Instead of adding a very large value that fades to zero over time,
just set the score to a value large enough to guarantee that this
network will get selected during a configurable time interval.

The default interval is 480 minutes (8 hours).

Bug: 151564543
Test: atest com.android.server.wifi
Change-Id: I2431415138adba1f16f86982f56275c4240ad81d