Clean up WifiTrafficPoller
Previously, WifiTrafficPoller only triggers callbacks
if the new data activity is different from the last
one.

However, this can cause some unexpected behaviors.
For example, if the data activity is constantly
DATA_ACTIVITY_INOUT i.e. there's always inbound &
outbound traffic, then if a new callback is registered,
it wouldn't be triggered until the data activity
changed, which could be a long time after.
But, the client would reasonably expect to be notified
for the first time soon after registration.

Modified WifiTrafficPoller to instead keep a
isFirstInvocation flag for each callback, and always
trigger the callback if it is the first invocation,
no matter if the data activity changed or not.

Also cleaned up this class.

Bug: 152633766
Test: atest WifiTrafficPollerTest
Change-Id: Ide4d040bf8602d9e8b6b9b3715d8d6e58618b408