[WifiTrackerLib] Make scanner run on single thread
Scanner should run on single thread to prevent race condition where
onStop() is called on the main thread while a msg is being processed on
the worker thread. This would result in a looping msg being sent after
the msg queue is cleared, which keeps the scan loop going even after
onStop().

Bug: 155425309
Test: build
Change-Id: Iecbab79b511bddf30515de27a60ff8eaf431d10e