WifiNative: Trigger disconnect on setMacAddress
This will ensure that we get the NETWORK_DISCONNECTION_EVENT with the
correct reason code. Changing the MAC address is only done when
connecting to a new network or on reboot. In both the cases,
triggering an explicit disconnect should be harmless (instead of
waiting for the implicit one if connected because of iface down).

Bug: 155899047
Test: Switched connection between networks and ensured that the
disconnect reason code is correct.

Change-Id: Icdf39ad71cb2872d09a769c85a6757fa2e0516ef