[WifiTrackerLib] Settings APP crashes after clicking a Passpoint2.0 Wi-Fi network twice
After the 1st click, WifiPickerTracker#getConnectedWifiEntry() will return
the passpoint2.0 OsuWifiEntry, however, NetworkDetailsTracker only supports
the key of a StandardWifiEntry or PasspointWifiEntry.

NetworkDetailsTracker throws below exception when it gets a key of OsuWifiEntry.

This change prevents a OsuWifiEntry to be returned by
WifiPickerTracker#getConnectedWifiEntry(). After the 1st click, the OsuWifiEntry
will disappear on WifiPicker and then it's online sign up screen shows.

Bug: 158055954
Test: manual visual
      Click a OsuWifiEntry twice before it's online sign up screen shows.
      After users go back to Wi-Fi picker from the online sign up screen,
      the result may be:

      1. If sign up failed, the OsuWifiEntry shows with the summary "Couldn't connect".
      2. If sign up success, there is a connected/connecting PasspointWifiEntry.

Change-Id: I89a16c51b001d16c1a5e2f4a8b0f0f942ce8ec72