Revert "Bug fix channel utilization during high traffic"
This reverts commit 2800851c885bd0f228e42a587c1db33da98f29ce.

Reason for revert: the merged CL has a wrong Merged-In which prevents the CL being auto-merged to other RVC branches including mainline-prod. Will re-submit a new CL with correct Merged-In so that it can be still auto-merged to other RVC branches but not into master.

Bug: 159866355
Test: atest com.android.server.wifi
Change-Id: I3fb0484ca365c13a547d371b69fe5548a39bdb99
Merged-In: I6e131f149d242e5dbf0b8b0087303e85ed428f9f