[WifiTrackerLib] Add null check to setOsuWifiEntry()
PasspointWifiEntry.setOsuWifiEntry() needs a null check for the passed
in OsuWifiEntry. This flow happens if the user is on the network details
page for a passpoint network and a corresponding OSU AP is also in
range. If the user exits the range of the OSU AP or scan results fail to
come in, then setOsuWifiEntry(null) will be called and cause an NPE.

Bug: 163328661
Test: atest WifiTrackerLibTests
Change-Id: I01790893a067489b671997a4fb4069ca4d3be8d3