[WifiTrackerLib] Handle isAutoJoinEnabled when no configs available
PasspointWifiEntry may have cases where mPasspointConfig and mWifiConfig
are both null, which can happen for a suggestion passpoint entry that
goes out of range, or the brief moment when a subscription passpoint
entry is forgotten but not yet removed. isAutoJoinEnabled needs to
default to false in case both configs are missing. In addition, to
prevent the wrong value from appearing for the out of range suggestion
case, keep a reference to the old WifiConfig when updating with null
scans.

Bug: 163084959
Test: atest WifiTrackerLibTests
Change-Id: I5eb40308cc01dc6e6f25b90828dd551f0cf4c633