[WifiTrackerLib] Add isDefaultNetwork boolean to WifiEntry
Added boolean isDefaultNetwork to WifiEntry.ConnectedInfo, which may be
retrieved through getConnectedInfo() for a connected WifiEntry. This
value will be dynamically updated as the default route changes to and
from Wifi. For a connected WifiEntry that is validated but not default,
the summary will display "Low Quality".

Bug: 163627176
Test: atest LocalWifiTrackerLibTests, manual verification of verbose log
indicating correct default route state when connecting to validated and
unvalidated networks.

Change-Id: I35bb3f57a5144311717165aa2be510365db11efb